pedido inser 2 tablas

+++ src/main/java/com/ues/model/DetallePedido.java
@@ -13,18 +13,18 @@ public class DetallePedido  implements java.io.Serializable {
      private int idDetallePedido;
      private Producto producto;
      private Pedido pedido;
-     private BigDecimal cantidad;
+     private int cantidad;
 
     public DetallePedido() {
     }
 
 	
-    public DetallePedido(int idDetallePedido, Pedido pedido, BigDecimal cantidad) {
+    public DetallePedido(int idDetallePedido, Pedido pedido, int cantidad) {
         this.idDetallePedido = idDetallePedido;
         this.pedido = pedido;
         this.cantidad = cantidad;
     }
-    public DetallePedido(int idDetallePedido, Producto producto, Pedido pedido, BigDecimal cantidad) {
+    public DetallePedido(int idDetallePedido, Producto producto, Pedido pedido, int cantidad) {
        this.idDetallePedido = idDetallePedido;
        this.producto = producto;
        this.pedido = pedido;
@@ -52,11 +52,11 @@ public class DetallePedido  implements java.io.Serializable {
     public void setPedido(Pedido pedido) {
         this.pedido = pedido;
     }
-    public BigDecimal getCantidad() {
+    public int getCantidad() {
         return this.cantidad;
     }
     
-    public void setCantidad(BigDecimal cantidad) {
+    public void setCantidad(int cantidad) {
         this.cantidad = cantidad;
     }
